Update mozilla components
This just updates to nightly versions and updates a few parts. During this iteration, m.facebook wouldn't reload (specifically reload, not app start, not going back, and not other subpages). It did eventually fix itself with the updates, but not sure what was happening. I've had very old builds work as intended, so I don't think it was a version issue alone